Prevent usage of Settings.h in headers
