+++ ci/jobs/scripts/check_style/check-settings-style
@@ -110,3 +110,23 @@ do
         echo "# In $line but it should be ${expected/$'\n'/ }"
     done
 done
+
+# Settings.h (and a few similar headers) is included in many parts of the
+# codebase, so any modifications to it trigger a large-scale recompilation.
+# Therefore, it is crucial to avoid unnecessary inclusion of Settings.h in
+# headers.
+SETTINGS_H_EXCLUDES=(
+    # Templated code
+    --exclude "$ROOT_PATH/src/Access/resolveSetting.h"
+    # TODO
+    --exclude "$ROOT_PATH/src/Client/ClientBase.h"
+    --exclude "$ROOT_PATH/src/Interpreters/AsynchronousInsertQueue.h"
+
+    # For functions we allow it for regular functions (due to lots of
+    # templates), but forbid it in interface (IFunction) part.
+    --exclude "$ROOT_PATH/src/Functions/*"
+    --include "$ROOT_PATH/src/Functions/IFunction*"
+)
+find $ROOT_PATH/src -name '*.h' -print0 | xargs -0 grep -P '#include[\s]*(<|")(Core/Settings.h)|(MergeTreeSettings.h)(>|")' "${SETTINGS_H_EXCLUDES[@]}" | \
+    grep . && echo '^ Too broad Settings.h or MergeTreeSettings.h usage'
+
