+++ src/IO/LimitReadBuffer.h
@@ -4,7 +4,6 @@
 #include <memory>
 #include <base/types.h>
 #include <IO/ReadBuffer.h>
-#include "Core/Settings.h"
 
 
 namespace DB
