+++ src/Storages/Distributed/DistributedAsyncInsertDirectoryQueue.cpp
@@ -12,6 +12,7 @@
 #include <IO/WriteBufferFromFile.h>
 #include <IO/ConnectionTimeouts.h>
 #include <Compression/CompressedReadBuffer.h>
+#include <Core/Settings.h>
 #include <Disks/IDisk.h>
 #include <Core/BackgroundSchedulePool.h>
 #include <Common/CurrentMetrics.h>
@@ -429,7 +430,7 @@ void DistributedAsyncInsertDirectoryQueue::processFile(std::string & file_path, 
             __PRETTY_FUNCTION__,
             storage.getContext()->getOpenTelemetrySpanLog());
 
-        Settings insert_settings = distributed_header.insert_settings;
+        Settings insert_settings = *distributed_header.insert_settings;
         insert_settings.applyChanges(settings_changes);
 
         auto timeouts = ConnectionTimeouts::getTCPTimeoutsWithFailover(insert_settings);
@@ -620,7 +621,7 @@ void DistributedAsyncInsertDirectoryQueue::processFilesWithBatching(bool force, 
             }
 
             BatchHeader batch_header(
-                std::move(distributed_header.insert_settings),
+                std::move(*distributed_header.insert_settings),
                 std::move(distributed_header.insert_query),
                 std::move(distributed_header.client_info),
                 std::move(header)
