+++ src/Storages/Kafka/IKafkaExceptionInfoSink.h
@@ -1,9 +1,6 @@
 #pragma once
 
 #include <memory>
-#include <cppkafka/cppkafka.h>
-#include <Core/Defines.h>
-#include <Core/Settings.h>
 
 namespace DB
 {
